Add react hot loader 3
