+++ package.json
@@ -80,16 +80,17 @@
     "lodash": "4.17.4",
     "material-ui": "0.17.1",
     "mocha": "3.2.0",
-    "node-sass": "6.0.2",
+    "node-sass": "4.5.2",
     "npm-run-all": "4.0.2",
     "object-assign": "4.1.1",
     "open": "0.0.5",
     "postcss-loader": "1.3.3",
-    "react": "15.4.2",
     "react-addons-test-utils": "15.4.2",
     "react-cookie": "0.4.8",
     "react-dom": "15.4.2",
+    "react-hot-loader": "3.0.0-beta.6",
     "react-redux": "5.0.3",
+    "react": "15.4.2",
     "react-router": "3.0.2",
     "react-router-redux": "4.0.8",
     "react-swipeable-views": "0.10.8",
